requirements.txt
cbor2==5.4.6
certifi==2023.7.22
charset-normalizer==3.2.0
idna==3.4
jsonpickle==3.0.2
luma.core==2.4.0
luma.led-matrix==1.7.0
paho-mqtt==1.6.1
Pillow==10.0.0
pyftdi==0.54.0
pyserial==3.5
python-dotenv==1.0.0
pyusb==1.2.1
requests==2.31.0
RPi.GPIO==0.7.1
smbus2==0.4.2
spidev==3.6
urllib3==2.0.4
